+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/react
 
+## 2.0.27
+
+### Patch Changes
+
+- Updated dependencies [ca40fac]
+  - ai@5.0.27
+
 ## 2.0.26
 
 ### Patch Changes